Add First Ever CSS Document
A css document was created and linked to both the homepage (index.html)
and caldo de camaron page( caldodecamaron.html). Using external css
method style was added to homepage and caldo de camaronpage. A purple
background given to the homepage and a green background to the caldopage.
Main headings (h1) were centered and customized using different classes.
In the caldo de camaronpage the image was centered using a <div> and
adding a class and aligning the div to the center. The font-size for
all 3 sections was increased to 18px. In the main page the caldo link
added to be part of an unordered list so in the future further links
to other recipes can be added.